Extract `function-to-string` tests from `sinon-test`.
Pulls out a suite into `test/util/core/funciton-to-string-test` for better organisation.
